+++ packages/kue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @idearium/kue
 
+## Unreleased
+
+### Changed
+
+-   Added `queue.constructCreateMethod`.
+
 ## v1.0.0-beta.1
 
 -   First version of the package.